GRF-240: move all ServiceApi related code into back directory. Put ServiceApi under infrastructure instead of root level. Move Interface in Domain.
